@@ -1,84 +1,64 @@
 ###
-# Blog settings
+# Page options, layouts, aliases and proxies
 ###
 
 Time.zone = "America/Los_Angeles"
 
-
-activate :blog do |blog|
-  # blog.prefix = "blog"
-  # blog.permalink = ":year/:month/:day/:title.html"
-  # blog.sources = ":year-:month-:day-:title.html"
-  blog.taglink = "/tags/:tag.html"
-  blog.layout = "post"
-  # blog.summary_separator = /(READMORE)/
-  # blog.summary_length = 250
-  # blog.year_link = ":year.html"
-  # blog.month_link = ":year/:month.html"
-  # blog.day_link = ":year/:month/:day.html"
-  # blog.default_extension = ".markdown"
-
-  blog.tag_template = "tag.html"
-  blog.calendar_template = "calendar.html"
-
-  blog.paginate = true
-  # blog.per_page = 10
-  # blog.page_link = "page/:num"
-end
-
-###
-# Compass
-###
-
-# Change Compass configuration
-# compass_config do |config|
-#   config.output_style = :compact
-# end
-
-###
-# Page options, layouts, aliases and proxies
-###
-
-page "feed.xml", :layout => false
-
-activate :syntax, wrap: true
-
-activate :directory_indexes
-
-set :markdown_engine, :redcarpet
-set :markdown, :layout_engine => :erb,
-               :fenced_code_blocks => true,
-               :tables => true,
-               :autolink => true,
-               :smartypants => true,
-               :with_toc_data => true
+# Per-page layout changes:
+#
+# With no layout
+page '/*.xml', layout: false
+page '/*.json', layout: false
+page '/*.txt', layout: false
+page "/feed.xml", layout: false
 
 ###
 # Helpers
 ###
 
-# Automatic image dimensions on image_tag helper
-# activate :automatic_image_sizes
+set :js_dir, 'js'
+set :images_dir, 'images'
+set :relative_links, true
+set :fonts_dir,  "fonts"
 
-# Methods defined in the helpers block are available in templates
-helpers do
+activate :syntax, wrap: true
+activate :directory_indexes
+
+set :markdown_engine, :redcarpet
+set :markdown, :layout_engine => :erb,
+  :fenced_code_blocks => true,
+  :tables => true,
+  :autolink => true,
+  :smartypants => true,
+  :with_toc_data => true
+
+
+activate :blog do |blog|
+  blog.taglink = "tags/{tag}.html"
+  blog.layout = "post"
+
+  blog.tag_template = "tag.html"
+  blog.calendar_template = "calendar.html"
+
+  blog.paginate = true
+  blog.per_page = 10
+  blog.page_link = "page/{num}"
 end
 
-set :css_dir, 'css'
+# Reload the browser automatically whenever files change
+configure :development do
+  activate :livereload
+end
 
-set :js_dir, 'js'
-
-set :images_dir, 'images'
-
-set :relative_links, true
+# Methods defined in the helpers block are available in templates
+# helpers do
+#   def some_helper
+#     "Helping"
+#   end
+# end
 
 # Build-specific configuration
 configure :build do
   activate :minify_css
   activate :minify_javascript
-  # Enable cache buster
-  # activate :cache_buster
-
-  # Use relative URLs
-  # activate :relative_assets
 end
